Fund operators from the start for testing purposes
